Clean up and update Gradle wrapper --> v4.6
Gradle version 4.6 includes native junit 5 support, so removed call to
Gradle JUnitPlatform plugin and configure test task instead.
Move all Block support code to /infrastructure/pagefactory
Remove Blocking demonstration 'tests'
Move footer tests to HomePageTests
