+++ poftutor/build.gradle
@@ -1,7 +1,9 @@
+import org.gradle.api.tasks.testing.junitplatform.JUnitPlatformOptions
+
 group 'tech.alexontest'
 version '1.0-SNAPSHOT'
 
-buildscript {
+/*buildscript {
     repositories {
         mavenCentral()
     }
@@ -25,7 +27,7 @@ junitPlatform {
         }
         includeClassNamePatterns '.*Test', '.*Tests'
     }
-}
+}*/
 
 apply plugin: 'java'
 sourceCompatibility = 1.8
@@ -53,11 +55,23 @@ dependencies {
     compile group: 'com.google.inject', name: 'guice', version: '4.1.0'
     compile group: 'com.google.code.gson', name: 'gson', version: '2.8.2'
     compile group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: junitVersion
-    compile group: 'org.seleniumhq.selenium', name: 'selenium-java', version: '3.8.1'
+    compile group: 'org.seleniumhq.selenium', name: 'selenium-java', version: '3.10.0'
     compile group: 'org.assertj', name: 'assertj-core', version: '3.9.0'
     compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.7'
 
     testCompile group: 'org.junit.jupiter', name: 'junit-jupiter-params', version: junitVersion
+    testCompile group: 'org.mockito', name: 'mockito-core', version: '2.7.22'
 
     testRuntime group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: junitVersion
+}
+
+test {
+    /*def options = new JUnitPlatformOptions()
+    options.setIncludeTags(["Framework"] as Set)*/
+    useJUnitPlatform()
+//    def excludeDirs = ["**/webdriver/*", "**/pagefactory/*"] as List
+//    exclude(["**/webdriver/*", "**/pagefactory/*"] as List)
+    options {
+        setExcludeTags(["Framework", "Blocking", "Navigation"] as Set)
+    }
 }
