+++ poftutor/src/main/java/tech/alexontest/poftutor/infrastructure/pagefactory/AbstractFoundBlock.java
@@ -1,15 +1,19 @@
-package tech.alexontest.poftutor.pageblocks;
+package tech.alexontest.poftutor.infrastructure.pagefactory;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.pagefactory.ElementLocatorFactory;
-import tech.alexontest.poftutor.infrastructure.pagefactory.BlockingElementLocatorFactory;
 
+/**
+ * Abstract super class to aid in the construction of blocks defined by a previously instantiated rootElement.
+ * Be cautious using these blocks for testing AJAX applications, the rootElement is prone to
+ * StaleElementReferenceExceptions if a DOM element above it in the tree is updated.
+ */
 public abstract class AbstractFoundBlock implements Block {
 
     private final WebElement rootElement;
 
-    AbstractFoundBlock(final WebElement rootElement) {
+    protected AbstractFoundBlock(final WebElement rootElement) {
         final ElementLocatorFactory locatorFactory = new BlockingElementLocatorFactory(rootElement);
         PageFactory.initElements(locatorFactory, this);
         this.rootElement = rootElement;
