+++ poftutor/src/main/java/tech/alexontest/poftutor/infrastructure/pagefactory/Block.java
@@ -1,0 +1,12 @@
+package tech.alexontest.poftutor.infrastructure.pagefactory;
+
+import org.openqa.selenium.WebElement;
+
+/**
+ * Interface for the support of PageBlock classes.
+ * Implementations should use the PageFactory to proxy fields relative to the WebElement rootElement field.
+ */
+@FunctionalInterface
+public interface Block {
+    WebElement getRootElement();
+}
