--- poftutor/src/main/java/tech/alexontest/poftutor/pageblocks/Block.java
@@ -1,8 +1,0 @@
-package tech.alexontest.poftutor.pageblocks;
-
-import org.openqa.selenium.WebElement;
-
-@FunctionalInterface
-public interface Block {
-    WebElement getRootElement();
-}
