+++ poftutor/src/main/java/tech/alexontest/poftutor/pageblocks/PostSummaryBlock.java
@@ -1,5 +1,7 @@
 package tech.alexontest.poftutor.pageblocks;
 
+import tech.alexontest.poftutor.infrastructure.pagefactory.Block;
+
 public interface PostSummaryBlock extends Block {
     String getPostTitle();
