+++ poftutor/src/main/java/tech/alexontest/poftutor/pageblocks/TagCloudWidgetBlock.java
@@ -2,6 +2,7 @@ package tech.alexontest.poftutor.pageblocks;
 
 import com.google.inject.ImplementedBy;
 import org.apache.commons.lang3.tuple.Pair;
+import tech.alexontest.poftutor.infrastructure.pagefactory.Block;
 
 import java.util.List;
